requirements.txt
anthropic>=0.40.0
groq>=0.11.0
google-generativeai>=0.8.0
supabase>=2.8.0
python-dotenv>=1.0.0
ollama>=0.4.0
